Finalize the Deck Planner step.
* Validation is implemented -- a deck configuration must be selected.
* Add displaying of error messages. Include an `error` prop in
DeckPlanner for the header to be displayed in bright red. (I'm not sure
what would be a more appropriate way to display this - perharps outline
it in bright red instead?)
* Change the data structure of `profiles` -- `numOfDecks` is now moved
to a new Object-type attribute called `config`, which also includes
other properties such as `perPlayer` (cards per player) and `spares`
(number of spare cards).