Revamp structure of `player`
* Things directly related to `player`, including `LevelDisplay`, have
been moved to `src/components/player`. This includes a defined
`PropType` for `player`, named `PlayerPropType`.
* The change from `player.active` being a `bool` to being a `number` has
been reverted. `player.active` will be back to being stored as a `bool`.
* Instead, the ID of the leader (since there can only be one every game)
will be stored under `player.leader`. When there is no defined leader
(i.e. in floating games, and before the first round of fixed games), it
is stored as `-1`.