Start work on fixed partnership profiles
* Extract some common styling options out to
'src/subpages/profiles/styles'
* Make `ScoreInput` function like a controlled component; you can change
the `value` prop passed to it for the component to display accordingly.
* Make the empty output for `ScoreInput` be `null` instead of
`undefined`, which is reserved for indeterminate input (i.e. the user is
still focused on the input).
* '/profile' will now display player details for fixed partnership
games. It will prompt the user to pick the winning bidder of the first
game. The rest behaves similarly to that of floating partnership
(with the exception of bidding). The level deltas are not shown;
instead, the role of each player in the next game will be shown.
* The row of the leader in player details, in all profiles (whether
fixed or floating partnership) will be highlighted.